requirements.txt
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2
google-api-core==1.10.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-storage==1.15.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.10
gunicorn==19.9.0
idna==2.8
numpy==1.16.2
Pillow==6.0.0
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.21.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
